Turn binary_heap::Kind into a sealed trait
